+++ traffic_portal/test/end_to_end/deliveryServices/pageData.js
@@ -19,9 +19,10 @@
 
 module.exports = function(){
 	this.moreBtn=element(by.name('moreBtn'));
+	this.createDSMenuItem=element(by.name('createDSMenuItem'));
 	this.viewStaticCapabilitiesMenuItem=element(by.css('a[ng-click*=viewStaticDnsEntries]'));
 	this.addCapabilityBtn=element(by.name('addCapabilityBtn'));
-	this.manageServersMenuItem=element(by.css('a[ng-click*=viewServers]'));
+	this.manageServersMenuItem=element(by.css('button[ng-click*=viewServers]'));
 	this.selectServersMenuItem=element(by.name('selectServersMenuItem'));
 	this.selectAllCB=element(by.id('selectAllCB'));
 	this.dsLink=element(by.name('dsLink'));